added support for auto learning
Added support for auto learning in itv_ascii and itv_words classes.
Which can be enabled / disabled from itv_config.h flag. When enabled,

1. during encryption, missing ascii / word (depending on class) will be
automatically added in itv table and appending to encrypted string.

2. during decryption, missing ascii / word is identified and added to
itv table with id (if present in encrypted string).

This way each itv table updates itself as it sees new data to encrypt.
On the decrypting end, it keep it self synced with itv table of remote
end.